SI-670 Remove OS specific impl from ETH if
