+++ src/drivers/meson.build
@@ -8,6 +8,7 @@ endif
 
 if host_machine.system() == 'linux'
 	csp_sources += files(['can/can_socketcan.c'])
+	csp_sources += files(['eth/eth_linux.c'])
 	csp_sources += files(['usart/usart_linux.c'])
 	csp_sources += files(['usart/usart_kiss.c'])
 elif host_machine.system() == 'windows'
